+++ sdk/src/main/java/com/stytch/sdk/api/StytchApi.kt
@@ -1,7 +1,12 @@
-package com.stytch.sdk
+package com.stytch.sdk.api
 
 import android.util.Base64
-import com.google.gson.annotations.SerializedName
+import com.stytch.sdk.Constants
+import com.stytch.sdk.ExceptionRecognizer
+import com.stytch.sdk.LoggerLocal
+import com.stytch.sdk.Stytch
+import com.stytch.sdk.StytchEnvironment
+import com.stytch.sdk.UnknownException
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import retrofit2.Call
@@ -9,58 +14,14 @@ import retrofit2.Callback
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.Body
-import retrofit2.http.POST
-import retrofit2.http.Path
 import java.net.InetAddress
 import java.net.NetworkInterface
 import java.util.Collections
 import java.util.concurrent.TimeUnit
 
-internal interface StytchService {
+internal class StytchApi {
 
-    @POST("magic_links/send_by_email")
-    fun sendMagicLink(
-        @Body request: SendMagicLinkRequest,
-    ): Call<SendMagicLingResponse>
-
-    @POST("magic_links/login_or_invite")
-    fun loginOrInvite(
-        @Body request: LoginOrInviteRequest,
-    ): Call<SendMagicLingResponse>
-
-    @POST("magic_links/login_or_create")
-    fun loginOrSignUp(
-        @Body request: LoginOrSignUpRequest,
-    ): Call<SendMagicLingResponse>
-
-    @POST("users")
-    fun createUser(
-        @Body request: CreateUserRequest,
-    ): Call<CreateUserResponse>
-
-    @POST("magic_links/{token}/authenticate")
-    fun verifyToken(
-        @Path("token") token: String,
-        @Body request: VerifyTokenRequest,
-    ): Call<VerifyTokenResponse>
-
-    @POST("emails/{email_id}/send_verification")
-    fun sendEmailVerification(
-        @Path("email_id") emailId: String,
-        @Body request: SendEmailVerificationRequest,
-    ): Call<SendEmailVerificationResponse>
-
-    @POST("users/{user_id}")
-    fun deleteUser(
-        @Path("user_id") userId: String,
-        @Body request: DeleteUserRequest,
-    ): Call<DeleteUserResponse>
-}
-
-internal class Api {
-
-    private val service: StytchService get() = ServiceGenerator.createService()
+    private val service: StytchApiService get() = ServiceGenerator.createService()
 
     fun sendMagicLink(
         email: String,
@@ -157,7 +118,7 @@ internal class Api {
     companion object {
         private const val TAG = "Api"
 
-        var instance = Api()
+        var instance = StytchApi()
 
         fun getIPAddress(useIPv4: Boolean): String? {
             try {
@@ -215,7 +176,7 @@ internal object ServiceGenerator {
     var apiBaseUrl = "https://api.stytch.com/v1/"
     var testApiBaseUrl = "https://test.stytch.com/v1/"
 
-    fun createService(): StytchService {
+    fun createService(): StytchApiService {
         val httpClient = OkHttpClient.Builder()
         httpClient
             .readTimeout(120L, TimeUnit.SECONDS)
@@ -236,112 +197,10 @@ internal object ServiceGenerator {
             .addConverterFactory(GsonConverterFactory.create())
             .client(client)
             .build()
-            .create(StytchService::class.java)
+            .create(StytchApiService::class.java)
     }
 }
 
-public data class StytchResult(
-    public val userId: String,
-    public val requestId: String,
-) {
-    override fun toString(): String {
-        return "StytchResult(userId='$userId', requestId='$requestId')"
-    }
-}
-
-public class BasicErrorResponse(
-    @SerializedName("status") public val status: Int,
-    @SerializedName("message") public val message: String?,
-    @SerializedName("error_type") public val error_type: String?,
-    @SerializedName("error_message") public val error_message: String?,
-    @SerializedName("error_url") public val error_url: String?,
-) {
-}
-
-public open class BasicResponse {
-}
-
-public class CreateUserResponse(
-    @SerializedName("request_id") public val request_id: String,
-    @SerializedName("user_id") public val user_id: String,
-    @SerializedName("email_id") public val email_id: String,
-) : BasicResponse() {
-}
-
-public class DeleteUserResponse : BasicResponse() {
-}
-
-public class SendEmailVerificationResponse(
-    @SerializedName("user_id") public val user_id: String,
-) : BasicResponse() {
-}
-
-public class SendMagicLingResponse(
-    @SerializedName("request_id") public val request_id: String,
-    @SerializedName("user_id") public val user_id: String,
-    @SerializedName("user_created") public val user_created: Boolean,
-    @SerializedName("email_id") public val email_id: String,
-) : BasicResponse() {
-}
-
-public class VerifyTokenResponse(
-    @SerializedName("request_id") public val request_id: String,
-    @SerializedName("user_id") public val user_id: String,
-) : BasicResponse() {
-}
-
-public open class BasicRequest {
-
-    @SerializedName("attributes")
-    public val attributes: Attributes = Attributes(
-        Api.getIPAddress(true),
-        System.getProperty("http.agent"),
-    )
-
-    public class Attributes(
-        @SerializedName("ip_address") public val ip_address: String?,
-        @SerializedName("user_agent") public val user_agent: String?,
-    )
-}
-
-public class CreateUserRequest(
-    @SerializedName("email") public val email: String,
-) : BasicRequest() {
-}
-
-public class DeleteUserRequest() : BasicRequest() {
-}
-
-public class LoginOrInviteRequest(
-    @SerializedName("email") public val email: String,
-    @SerializedName("login_magic_link_url") public val login_magic_link_url: String,
-    @SerializedName("invite_magic_link_url") public val invite_magic_link_url: String,
-    @SerializedName("login_expiration_minutes") public val login_expiration_minutes: Long,
-    @SerializedName("invite_expiration_minutes") public val invite_expiration_minutes: Long,
-) : BasicRequest() {
-}
-
-public class LoginOrSignUpRequest(
-    @SerializedName("email") public val email: String,
-    @SerializedName("login_magic_link_url") public val login_magic_link_url: String,
-    @SerializedName("signup_magic_link_url") public val signup_magic_link_url: String,
-    @SerializedName("login_expiration_minutes") public val login_expiration_minutes: Long,
-    @SerializedName("signup_expiration_minutes") public val signup_expiration_minutes: Long,
-) : BasicRequest() {
-}
-
-public class SendEmailVerificationRequest() : BasicRequest() {
-}
-
-public class SendMagicLinkRequest(
-    @SerializedName("email") public val email: String,
-    @SerializedName("magic_link_url") public val magic_link_url: String,
-    @SerializedName("expiration_minutes") public val expiration_minutes: Long,
-) : BasicRequest() {
-}
-
-public class VerifyTokenRequest(
-    @SerializedName("ip_match_required") public val ip_match_required: Boolean? = null,
-    @SerializedName("user_agent_match_required") public val user_agent_match_required: Boolean? = null,
-) : BasicRequest() {
+internal fun String.deepLink(): String {
+    return "${Stytch.instance.config?.deepLinkScheme}://${Stytch.instance.config?.deepLinkHost}/$this"
 }
