+++ sdk/src/main/java/com/stytch/sdk/ui/view/StytchTextView.kt
@@ -1,0 +1,21 @@
+package com.stytch.sdk.ui.view
+
+import android.content.Context
+import android.util.AttributeSet
+import android.util.TypedValue
+import androidx.appcompat.widget.AppCompatTextView
+import com.stytch.sdk.StytchTextStyle
+
+internal class StytchTextView constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0,
+) : AppCompatTextView(context, attrs, defStyleAttr) {
+
+    fun setCustomization(style: StytchTextStyle) {
+        setTextColor(style.color.getColor(context))
+        typeface = style.font
+        setTextSize(TypedValue.COMPLEX_UNIT_PX, style.size)
+    }
+
+}
